% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{get_feature_names}
\alias{get_feature_names}
\title{Get names of features of each dictionaries supported.}
\usage{
get_feature_names(dic = c("ipa", "uni", "ko"))
}
\arguments{
\item{dic}{The name of dictionary.}
}
\value{
character vector.
}
\description{
Get names of features of each dictionaries supported.
}
