% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kagome.R
\name{kagome}
\alias{kagome}
\title{Japanese morphological analyzer}
\usage{
kagome(text, ...)
}
\arguments{
\item{text}{character vector}

\item{...}{all other args are passed to \code{jsonlite::fromJSON}}
}
\value{
list
}
\description{
Call kagome tokenizer.
}
