% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{Pack output of kagome}
\usage{
pack(obj, ...)
}
\arguments{
\item{obj}{Object}

\item{...}{Other arguments}
}
\value{
data.frame
}
\description{
Alias of \code{RcppKagome::pack_list} and \code{RcppKagome::pack_df}
}
\seealso{
Other pack-fn: 
\code{\link{pack_df}()},
\code{\link{pack_list}()}
}
\concept{pack-fn}
