% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack_df}
\alias{pack_df}
\title{Pack prettified output}
\usage{
pack_df(df, pull = "token", .collapse = " ")
}
\arguments{
\item{df}{Output of \code{RcppKagome::prettify}.}

\item{pull}{Column name to be packed into data.frame. Default value is \code{token}.}

\item{.collapse}{This argument is passed to \code{stringi::stri_c()}.}
}
\value{
data.frame.
}
\description{
Pack prettified output
}
\seealso{
Other pack-fn: 
\code{\link{pack_list}()},
\code{\link{pack}()}
}
\concept{pack-fn}
