% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack_list}
\alias{pack_list}
\title{Pack ouput of kagome}
\usage{
pack_list(list, .collapse = " ")
}
\arguments{
\item{list}{Output of \code{RcppKagome::kagome}.}

\item{.collapse}{This argument is passed to \code{stringi::stri_c()}.}
}
\value{
data.frame.
}
\description{
Pack ouput of kagome
}
\seealso{
Other pack-fn: 
\code{\link{pack_df}()},
\code{\link{pack}()}
}
\concept{pack-fn}
