% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify}
\alias{prettify}
\title{Prettify kagome output}
\usage{
prettify(list, col_names = get_feature_names("ipa"))
}
\arguments{
\item{list}{Output of \code{RcppKagome::kagome}.}

\item{col_names}{Column names of output data.frame.}
}
\value{
data.frame.
}
\description{
Prettify kagome output
}
