% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kagomer-rexeports.R
\name{queue}
\alias{queue}
\title{Creates asynchronous requests}
\usage{
queue(params, url = paste(Sys.getenv("KAGOME_URL"), "tokenize", sep = "/"))
}
\arguments{
\item{params}{List of json string.}

\item{url}{An URL of Kagome server API.}
}
\value{
returns a list of json strings and asynchronous function.
}
\description{
Creates asynchronous requests
}
