% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppKagome.R
\docType{package}
\name{RcppKagome}
\alias{RcppKagome}
\alias{RcppKagome-package}
\title{RcppKagome: 'Rcpp' Interface to Kagome}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Interface to ikawaha/Kagome; Self-contained Japanese morphological analyzer
    written in pure Go.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/RcppKagome}
  \item Report bugs at \url{https://github.com/paithiov909/RcppKagome/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
\keyword{internal}
