% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kagomer-rexeports.R
\name{serialize}
\alias{serialize}
\title{Create json data}
\usage{
serialize(sentences, mode = "normal", dict = "ipa")
}
\arguments{
\item{sentences}{Character to be analyzed.}

\item{mode}{One of `normal`, `search` or `extended`.}

\item{dict}{One of `ipa` (the IPA-dic) or `uni` (the Uni-dic).}
}
\value{
a list of json strings.
}
\description{
Create json data
}
