% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_segments}
\alias{split_segments}
\title{Go implementation of TinySegmenter}
\usage{
split_segments(text, ...)
}
\arguments{
\item{text}{Character vector.}

\item{...}{All other args are passed to \code{jsonlite::fromJSON}.}
}
\value{
list.
}
\description{
Call TinySegmenter.
}
\details{
This function is a wrapper of \code{nyarla/go-japanese-segmenter}.
The \code{go-japanese-segmenter} library contians dictionary data
and dictionary code from the original \code{TinySegmenter.js} library written by Taku Kudo.
For more details, please look a glance at the license notice from 'See Also' section.
}
\seealso{
\url{https://github.com/nyarla/go-japanese-segmenter#licenses}
}
