% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_sentences}
\alias{split_sentences}
\title{Japanese sentence splitter}
\usage{
split_sentences(text, ...)
}
\arguments{
\item{text}{Character vector.}

\item{...}{All other arguments are passed to \code{jsonlite::fromJSON}.}
}
\value{
list.
}
\description{
Call sentence splitter.
}
