% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tokenize}
\alias{tokenize}
\title{Trigger kagome tokenizer}
\arguments{
\item{text}{character vector}
}
\value{
res character scalar (JSON string)
}
\description{
For internal use. The argument should be UTF8 encoded. This function just
returns an UTF8-encoded json as character scalar.
}
\keyword{internal}
