% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tokenize_morphemes}
\alias{tokenize_morphemes}
\title{Trigger kagome tokenizer}
\arguments{
\item{text}{Character vector.}
}
\value{
Character vector (JSON strings).
}
\description{
For internal use. The argument should be UTF8 encoded. This function just
returns UTF8-encoded json strings as a character vector.
}
\keyword{internal}
