% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tokenize_segments}
\alias{tokenize_segments}
\title{Tiny Segmenter}
\arguments{
\item{text}{Character vector.}
}
\value{
Character vector.
}
\description{
For internal use. The argument should be UTF8 encoded.
}
\keyword{internal}
