% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tokenize_sentences}
\alias{tokenize_sentences}
\title{Split sentence}
\arguments{
\item{text}{Character vector.}
}
\value{
res List.
}
\description{
For internal use. The argument should be UTF8 encoded.
}
\keyword{internal}
