% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kagome.R
\name{kagome}
\alias{kagome}
\title{Japanese morphological analyzer
Calls kagome tokenizer.}
\usage{
kagome(text, ...)
}
\arguments{
\item{text}{Character vector.}

\item{...}{All other arguments are passed to \code{jsonlite::fromJSON}.}
}
\value{
list.
}
\description{
Japanese morphological analyzer
Calls kagome tokenizer.
}
