% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kagomer-re-exports.R
\name{kick}
\alias{kick}
\title{Kick requests}
\usage{
kick(requests, keep = c("surface"), .skip_enc_reset = FALSE)
}
\arguments{
\item{requests}{List that comes out from \code{queue}.}

\item{keep}{Column names to keep in the results.}

\item{.skip_enc_reset}{Logical. If true, leaves encoding of results as environment native.}
}
\value{
data.frame.
}
\description{
Kick requests
}
