% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack_list}
\alias{pack_list}
\title{Packs ouput of kagome}
\usage{
pack_list(list, .collapse = " ")
}
\arguments{
\item{list}{Output of \code{RcppKagome::kagome}.}

\item{.collapse}{This argument will be passed to \code{stringr::str_c()}.}
}
\value{
data.frame
}
\description{
Packs ouput of kagome
}
