% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify}
\alias{prettify}
\title{Prettify kagome output}
\usage{
prettify(list)
}
\arguments{
\item{list}{Output of \code{RcppKagome::kagome}.}
}
\value{
data.frame.
}
\description{
Prettify kagome output
}
