% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kagomer-re-exports.R
\name{queue}
\alias{queue}
\title{Create asynchronous requests}
\usage{
queue(params, url = paste(Sys.getenv("KAGOME_URL"), "tokenize", sep = "/"))
}
\arguments{
\item{params}{List of json string.}

\item{url}{An URL of Kagome server.}
}
\value{
a list of json strings and asynchronous function.
}
\description{
Create asynchronous requests
}
