% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tokenize_morphemes}
\alias{tokenize_morphemes}
\title{Triggers kagome tokenizer}
\arguments{
\item{text}{Character vector.}
}
\value{
res Character scalar (JSON string).
}
\description{
For internal use. The argument should be UTF8 encoded. This function just
returns an UTF8-encoded json as a character scalar.
}
\keyword{internal}
